+++ examples/certificates.py
@@ -1,0 +1,41 @@
+###
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+##
+
+from simplivity.ovc_client import OVC
+
+import pprint
+
+pp = pprint.PrettyPrinter(indent=4)
+
+config = {
+    "ip": "<ovc_ip>",
+    "credentials": {
+        "username": "<username>",
+        "password": "<password>"
+    }
+}
+
+ovc = OVC(config)
+certificates = ovc.certificates
+
+print("\n\nget_all with default params")
+all_certificates = certificates.get_all()
+count = len(all_certificates)
+for certificate in all_certificates:
+    print(f"{certificate}")
+    print(f"{pp.pformat(certificate.data)} \n")
+
+print(f"Total number of policies : {count}")